requirements-dev.txt
codecov==2.0.15
pytest-cov==2.6.1
pytest-aiohttp==0.3.0
